AniTalker-kit/AniTalker-judge/requirements.txt
# python==3.9.21
imageio[ffmpeg]==2.36.1
torch==2.0.1 
torchvision==0.15.2 
torchaudio==2.0.2
pytorch_msssim==1.0.0
pillow==11.0.0
opencv-python-headless==4.10.0.84
pyiqa==0.1.13
scipy==1.13.1
ipykernel==6.29.5
ipywidgets==8.1.5
numpy==1.23.0